Some small bug fixes on overwolf.d.ts file
The namespace "paths" had some mistakes in the code like some const and propertyName were without a space like constpropertyName. But, even with that fixed, it throws an error because of the "," comma separation, and on closer look, I saw that the namespace "paths" probably should be an interface. If that's the case then there is no need for const, and in the case of const, it needs to be separated with ";" semicolon. I had one more issue on line 6010 but I am not sure what the err was. 

This all happened while I tried to install the types with Angular 10 using TS. Thanks.